Switch to dynamic linking with `liblightning`
It appears from rust-lang/rust#35061 and from [some build failures][1]
that position-independent executables are being built (see "-pie" flag
in [1]'s build output), which would require the objects in
`liblightning.a` to be built with `-fPIC` on x86_64, which is not
desirable.

Static linking would be more desirable in the long term, since it could
enable extra optimizations (`-flto` particularly) and generate better
code, but for now, switch to dynamic linking until a portable way
becomes available to ensure linking succeeds statically.

[1]: https://travis-ci.org/github/kulp/lightning-sys/jobs/685111507