Remove implied dependency on `wget`
Move downloading and unpacking of the GNU Lightning release into Rust-only
code, removing an implied dependency on `wget`, which is not available on all
platforms.

It is unfortunate that `reqwest` and all its baggage are the most obvious way
to download a file. In an effort to trim the dependency tree, default features
are specifically not included for `reqwest`, which therefore implies that
download will occur over HTTP and not HTTPS.