Translate doctests into explicit tests
Ideally this would not be necessary, but gettin Tarpaulin to compute coverage
for doctests (xd009642/tarpaulin#13) is supported, but trickier than it first
appears. For me, the following command appears to hang during the compilation
of a transitively-required crate:

    cargo tarpaulin --verbose --run-types Tests,Doctests

so as a workaround to institute coverage, the doctests are duplicated as
explicit tests.